gggg

+++ _print_functions.c
@@ -75,9 +75,9 @@ int print_non_printable(char *s)
  * @n: n.
  * Return: 1.
  */
-int print_number(int n)
+int print_number(long int n)
 {
-	unsigned int x;
+	unsigned long int x;
 	int printed_characters;
 
 	printed_characters = 0;
